requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PDF Processing
PyMuPDF==1.23.14
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# AI/ML
langchain>=0.1.0,<0.2.0
# langchain-openai==0.0.5
# openai>=1.10.0,<2.0.0
langchain-google-genai>=0.0.5
google-generativeai>=0.3.2
langchain-core>=0.1.7,<0.4.0
langsmith>=0.0.77
langgraph==0.0.20

# Data Models
pydantic>=2.5.2
numpy==1.24.3

# Utilities
python-dotenv==1.0.0